Share test dependencies using gradle testFixtures. Move test code to textFixtures. Upgrade to gradle 8.12
