+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/util/Utils.kt
@@ -1,6 +1,9 @@
 package org.cryptobiotic.rlauxe.util
 
+import java.security.SecureRandom
 import kotlin.math.abs
+
+val secureRandom = SecureRandom.getInstanceStrong()!!
 
 fun doubleIsClose(a: Double, b: Double, rtol: Double=1.0e-5, atol:Double=1.0e-8): Boolean {
     //    For finite values, isclose uses the following equation to test whether
@@ -17,6 +20,9 @@ fun listToMap(vararg names: String): Map<String, Int> {
 fun listToMap(names: List<String>): Map<String, Int> {
     return names.mapIndexed { idx, value -> value to idx }.toMap()
 }
+
+fun margin2mean(margin: Double) = (margin + 1.0) / 2.0
+fun mean2margin(mean: Double) = 2.0 * mean - 1.0
 
 fun df(d: Double) = "%6.4f".format(d)
 fun dfn(d: Double, n: Int) = "%${n+2}.${n}f".format(d)
