+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/core/TestAssertions.kt
@@ -1,6 +1,7 @@
 package org.cryptobiotic.rlauxe.core
 
 import org.cryptobiotic.rlauxe.doublePrecision
+import org.cryptobiotic.rlauxe.sampling.makeCvrsByExactCount
 import org.cryptobiotic.rlauxe.util.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
