+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/core/TestProbRatios.kt
@@ -2,7 +2,6 @@ package org.cryptobiotic.rlauxe.core
 
 import org.cryptobiotic.rlauxe.doublePrecision
 import kotlin.test.Test
-import java.lang.RuntimeException
 import kotlin.random.Random
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
