+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/sampling/TestComparisonWithErrors.kt
@@ -1,7 +1,6 @@
 package org.cryptobiotic.rlauxe.sampling
 
 import org.cryptobiotic.rlauxe.core.*
-import org.cryptobiotic.rlauxe.util.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
