+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/shangrla/TestShrinkTrunc.kt
@@ -3,8 +3,8 @@ package org.cryptobiotic.rlauxe.shangrla
 import org.cryptobiotic.rlauxe.core.PrevSamples
 import org.cryptobiotic.rlauxe.core.TruncShrinkage
 import org.cryptobiotic.rlauxe.core.populationMeanIfH0
-import org.cryptobiotic.rlauxe.doubleIsClose
 import org.cryptobiotic.rlauxe.doublesAreClose
+import org.cryptobiotic.rlauxe.util.doubleIsClose
 import kotlin.test.Test
 import kotlin.test.assertTrue
