+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/util/TestCvrs.kt
@@ -4,6 +4,8 @@ import org.cryptobiotic.rlauxe.core.Contest
 import org.cryptobiotic.rlauxe.core.ContestInfo
 import org.cryptobiotic.rlauxe.core.ContestUnderAudit
 import org.cryptobiotic.rlauxe.core.SocialChoiceFunction
+import org.cryptobiotic.rlauxe.sampling.makeContestsFromCvrs
+import org.cryptobiotic.rlauxe.sampling.makeCvrsByExactMean
 import org.cryptobiotic.rlauxe.workflow.makeContestUAFromCvrs
 import kotlin.test.Test
 import kotlin.test.assertEquals
