+++ core/src/testFixtures/kotlin/org/cryptobiotic/rlauxe/sampling/Utils.kt
@@ -1,0 +1,18 @@
+package org.cryptobiotic.rlauxe.sampling
+
+import kotlin.random.Random
+
+///////////////////////
+//// DoubleArrays
+fun randomPermute(samples : DoubleArray): DoubleArray {
+    val n = samples.size
+    val permutedIndex = MutableList(n) { it }
+    permutedIndex.shuffle(Random)
+    return DoubleArray(n) { samples[permutedIndex[it]] }
+}
+
+
+
+
+
+
