+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,8 +1,8 @@
-#Mon Aug 26 14:42:02 MDT 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.6-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.12-bin.zip
 kotest.framework.classpath.scanning.autoscan.disable=true
-kotest.framework.parallelism=24
+networkTimeout=10000
+validateDistributionUrl=true
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
