+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/comparison/Failures.kt
@@ -1,7 +1,7 @@
 package org.cryptobiotic.rlauxe.comparison
 
 import org.cryptobiotic.rlauxe.core.Cvr
-import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
+import org.cryptobiotic.rlauxe.sampling.makeCvrsByExactMean
 import org.cryptobiotic.rlauxe.rlaplots.SRTcsvWriter
 import org.cryptobiotic.rlauxe.sim.RepeatedTaskRunner
 import org.cryptobiotic.rlauxe.util.mean2margin
